Move elasticsearch methods to own folder
