+++ elasticsearch/server/methods.js
@@ -1,6 +1,8 @@
 import { Meteor } from 'meteor/meteor';
-import { Settings } from '/settings/collection';
 import ElasticSearch from 'elasticsearch';
+// Collection import
+import { Proxies } from '/proxies/collection';
+
 
 Meteor.methods({
   getElasticSearchData (opts) {
@@ -21,5 +23,28 @@ Meteor.methods({
 
       return false;
     }
+  },
+  elasticsearchIsDefined () {
+    // TODO: multi-proxy support
+    const proxy = Proxies.findOne();
+
+    if (proxy) {
+      const elasticsearch = proxy.apiUmbrella.elasticsearch;
+
+      // Return true or false, depending on whether elasticsearch is defined
+      return (elasticsearch);
+    }
+
+    return false;
+  },
+  getElasticsearchUrl () {
+    if (Meteor.call('elasticsearchIsDefined')) {
+      // TODO: multi-proxy support
+      const elasticsearch = Proxies.findOne().apiUmbrella.elasticsearch;
+
+      return elasticsearch;
+    }
+
+    throw new Meteor.Error('Elasticsearch is not defined');
   },
 });
